requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
python-dotenv==1.1.0
instagrapi==2.1.5
mysqlclient==2.2.7
mysql-connector-python==9.3.0